Add file-name-without-extension
Often times you need to get just the file name without the extension,
or the file name with a specific extension removed if it is present.
Not including the parent path in the result since the user can prepend
that if needed and it's not at all a complicated task.

(path (parent my-path) (file-name-without-extension my-path))