feat: add with-temp-dir macro
The macro takes a sequence of binding-name, options pairs.

The body is evaluated with binding-name bound to a java.io.File path to
a temporary directory, created by passing options to create-temp-dir.

The temporary directories are deleted with delete-tree in the reverse
of the order in which they were declared.

Options are as for create-temp-dir.

Adds an exported clj-kondo config for the macro.