docstrings: use wikilinks instead of backticks
This allows us to be deliberate with var links and to not automatically,
for example, link a `path` arg to the `path` var.

Also nice for cljdoc which only links to vars with wikilinks.

I notice the quickdoc toc summary uses the wikilink syntax which is
probably not ideal. I can swing back to quickdoc to address that later
if you'd like.

Contributes to #160